requirements.txt
discord.py
pandas
numpy<2
pandas-ta
yfinance
SQLAlchemy
alembic
apscheduler
matplotlib
python-dotenv
psycopg2-binary
setuptools

